Add `DelegatingServiceDiscoverer` to client-api
Motivation:

`DelegatingServiceDiscoverer` is a helpful utility to reduce
boilerplate when users need to implement a filter for
`ServiceDiscoverer`. We provide "delegate" classes for other use-cases,
would be helpful to have one for `ServiceDiscoverer` too.

Modifications:
- Add `io.servicetalk.client.api.DelegatingServiceDiscoverer`;
- Use new utility instead of internal variant inside
`DefaultSingleAddressHttpClientBuilder`;

Result:

Easier for users to implement a filter for `ServiceDiscoverer`.
